activate via topnav trigger, dismiss via esc, enter animation
